+++ obfuscated/src/com/fs/starfarer/combat/entities/ship/trackers/BeamChargeTracker.java
@@ -1,0 +1,9 @@
+package com.fs.starfarer.combat.entities.ship.trackers;
+
+// OBFUSCATED
+public class BeamChargeTracker {
+    // OBFUSCATED
+    public BeamWeaponState beamChargeTracker_getState() {
+        return null;
+    }
+}