+++ obfuscated/src/com/fs/starfarer/combat/entities/ship/trackers/BeamWeaponState.java
@@ -1,0 +1,4 @@
+package com.fs.starfarer.combat.entities.ship.trackers;
+
+// OBFUSCATED
+public enum BeamWeaponState {}