+++ obfuscated/src/com/fs/starfarer/combat/entities/ship/weapons/BeamWeapon.java
@@ -1,0 +1,11 @@
+package com.fs.starfarer.combat.entities.ship.weapons;
+
+import com.fs.starfarer.combat.entities.ship.trackers.BeamChargeTracker;
+
+// OBFUSCATED
+public class BeamWeapon {
+    // UNOBFUSCATED
+    public BeamChargeTracker getChargeTracker() {
+        return null;
+    }
+}